requirements.txt
poetry
hydra-core
torch
torchvision
matplotlib